Fixed from_future
